70834: Intermediate commit
